FeatherMyth
A Legend of Zelda inspired game of navigating room, killing monsters. 

Prompt: 
Now, in the same style, make a Zelda type clone with 3 rooms Start -> Right -> Up. 

The game style is pixel art. 
Legend of Zelda (NES) type view.
There is a sword to pickup in the first room. 
There is a boomerang to pickup in the second room. 
The first room is a green field - rain is on. The second room is a grey room - rain is off. Third room is a throne room with a dark red carpet - rain is on, screen flashes from lightning (white) from time to time.
Protagonist has 3 health boxes (30 life points) displayed as 3 boxes red, side by side. They turn black when protagonist is hit. When hit by enemy, protagonist is affected by knockback and loses 1 health (10 life points). At 0 health game resets and send you back to room 1. 
Third room has a 1 health pickup . 
Sword is close range attack. It kills enemy in one hit (damage 10) 
boomerang is longer ranged attack, it goes through enemies hurting them and returns to protagonist (damage 5). 
All monsters have 10 life points, unless mentioned. In the first room, there is no enemy. In the second room, there are 3 purple bat monsters that fly around the room on a circular route, changing direction when they meet a wall. Making a complex dance that seems a bit random. The third room has a monster (centipede) made out of 4 segments. Each hit on the monster destroys one segment. The centipede only moves in straight lines, but moves faster than the protagonist. The monster must move at least 1/4 of the room before changing direction left or right. 
Enemies drop a gem shaped Rupee (money) when killed. This is a pickup. Count the number of Rupees collected. 
The exit of a room is indicated on the wall with a rectangle that is black. Walking into it brings you into the next room (zelda scroll style - protagonist disappears, room scrolls into view, protagonist appears, room starts). 

Since it took a few iterations I added: 
When centipede is killed, rain and thunder stops and a halo of yellow light illuminates the room. 
Centipede fight too easy, it's not aiming to hit the hero - try to chase the hero. also it's allowed to continue straight too instead of only left right turns. Maybe diminish the minimum path it needs to do a little. Also make it 20% faster for each segment it loses.
When hit by the boomerang the monsters should flash one and freeze .3 seconds. Also the boomerang is allowed to hit them twice (one on the way there and once on the way back)
The protagonist must be placed on the opposite side of the room where he exited the last room, so as to match the entrance.