Prompt TetrOut
Write a single page javascript game that starts out as a tetris, but the bricks cannot be eliminated by making a row. Once the bricks fill up to the top, the game continues like breakout, with a ball sent in from the bottom that bounces around and breaks the tetromino squares that make up the bricks. A paddle appears at the bottom to make it bounce. Of course this means that in the tetris part, there's a space below that the bricks cannot occupy, that is represented by a thin red line - the paddle appears below. The paddle can be moved by clicking and dragging the mouse left / right - or by using the keyboard keys. Every time a row is cleared in breakout mode, an extra ball gets added to the game. When all tetra are destroyed, the game starts again stacking tetra. HTML+CSS+JS game in a single page. Add some flourish to make it look cool and some effects. The score increments and the speed should slowly ramp every time a screen is filled / cycled (that's the level counter).